Allow a blank GCloud service account to be selected.
The default credential selection ("-none-") implies that there should be no
authentication performed. Normally this would be unusual because most gcloud
commands will require auth (but not all!).

Since there is also the option to execute a gcloud command with auth, that
also gives another chance to select a different service account. And if the
gcloud command does require auth, it will fail with:

    ERROR: (gcloud.the.command.that.ran)
    You do not currently have an active account selected.

which should be sufficient to indicate that they need to select credentials.